Make flock mechanism to work on NFS environments (bsc#1240666)
Since Linux 2.6.12, NFS clients support flock() locks by emulating
them as fcntl(2) byte-range locks on the entire file.  This means
that fcntl(2) and flock() locks do interact with one another over
NFS.  It also means that in order to place an exclusive lock, the
file must be opened for writing.

https://man7.org/linux/man-pages/man2/flock.2.html